+++ src/Playgo.Domain/Enums/Enums.cs
@@ -20,7 +20,8 @@ public enum VideoQuality
 {
     SD = 1,
     HD = 2,
-    FullHD = 3
+    FullHD = 3,
+    UHD4K = 4
 }
 
 public enum UserRole